refactor[#59] : change dto directory path of member and post (domain => application)
