+++ src/main/java/com/itaxi/server/member/application/dto/MemberInfo.java
@@ -1,4 +1,4 @@
-package com.itaxi.server.member.domain.dto;
+package com.itaxi.server.member.application.dto;
 
 public interface MemberInfo {
     Long getId();
