+++ src/main/java/com/itaxi/server/member/domain/repository/MemberRepository.java
@@ -1,8 +1,8 @@
 package com.itaxi.server.member.domain.repository;
 
 import com.itaxi.server.member.domain.Member;
-import com.itaxi.server.member.domain.dto.LoginResponse;
-import com.itaxi.server.member.domain.dto.MemberInfo;
+import com.itaxi.server.member.application.dto.LoginResponse;
+import com.itaxi.server.member.application.dto.MemberInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
