+++ src/main/java/com/itaxi/server/post/presentation/response/PostInfoResponse.java
@@ -1,7 +1,7 @@
 package com.itaxi.server.post.presentation.response;
 
 import com.itaxi.server.place.application.PlaceResponse;
-import com.itaxi.server.post.domain.dto.JoinerInfo;
+import com.itaxi.server.post.application.dto.JoinerInfo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
